WIP moved all packages and enabled cross compiling to 2.11 and 2.12. Docs currently failing
