+++ build.sbt
@@ -10,7 +10,8 @@ val gh   = GitHubSettings("com.fortysevendeg", "freestyle", "47 Degrees", apache
 val vAll = Versions(versions, libraries, scalacPlugins)
 
 lazy val commonSettings = Seq(
-  scalaVersion in ThisBuild := "2.11.8",
+  scalaVersion in ThisBuild := "2.12.0",
+  crossScalaVersions in ThisBuild := Seq("2.11.8", "2.12.0"),
   scalaOrganization in ThisBuild := "org.typelevel",
   addCompilerPlugin("tryp" %% "splain" % "0.1.11"),
     resolvers ++= Seq(
@@ -39,7 +40,6 @@ lazy val commonSettings = Seq(
     "-unchecked",
     //"-Xfatal-warnings",
     //"-Xlint",
-    "-Yinline-warnings",
     "-Yno-adapted-args",
     "-Ywarn-dead-code",
     "-Ywarn-numeric-widen",
@@ -89,7 +89,7 @@ lazy val freestyle = (crossProject in file("freestyle")).
   settings(
     libraryDependencies ++= Seq(
       "org.typelevel" %%% "cats-free" % "0.8.1",
-      "org.scala-lang" % "scala-reflect" % "2.11.8"
+      "org.scala-lang" % "scala-reflect" % scalaVersion.value
     )
   ).
   jsSettings(sharedJsSettings: _*)
@@ -158,7 +158,7 @@ lazy val freestyleConfig = (crossProject in file("freestyle-config")).
   ).
   settings(
     libraryDependencies ++= Seq(
-      "eu.unicredit" %%% "shocon" % "0.1.4",
+      "eu.unicredit" %%% "shocon" % "0.1.7",
       "org.scalatest" %%% "scalatest" % "3.0.0" % "test"
     )
   )
@@ -191,7 +191,7 @@ lazy val freestyleLogging = (crossProject in file("freestyle-logging")).
     libraryDependencies += "org.scalatest" %%% "scalatest" % "3.0.0" % "test"
   ).
   jvmSettings(
-    libraryDependencies += "io.verizon.journal" %% "core" % "2.3.16"
+    libraryDependencies += "io.verizon.journal" %% "core" % "3.0.18"
   ).
   jsSettings(
     libraryDependencies += "biz.enef" %%% "slogging" % "0.5.2"
@@ -208,7 +208,7 @@ lazy val tests = (project in file("tests")).
   settings(noPublishSettings: _*).
   settings(
     libraryDependencies ++= Seq(
-      "org.scala-lang" % "scala-reflect" % "2.11.8",
+      "org.scala-lang" % "scala-reflect" % scalaVersion.value,
       "org.scalatest" %%% "scalatest" % "3.0.0" % "test"
     )
   )
@@ -220,6 +220,7 @@ lazy val docs = (project in file("docs")).
   settings(noPublishSettings: _*).
   settings(
     name := "docs",
-    description := "freestyle docs"
+    description := "freestyle docs",
+    addCompilerPlugin("org.scalamacros" % "paradise" % "2.1.0" cross CrossVersion.full)
   ).
   enablePlugins(MicrositesPlugin)
