+++ docs/src/main/tut/docs/algebras.md
@@ -16,7 +16,7 @@ without worrying about how they will be implemented.
 A trait or abstract class annotated with `@free` is all you need to create your first algebra with Freestyle.
 
 ```tut:silent
-import io.freestyle._
+import freestyle._
 
 case class User(id: Long, name: String)
