+++ freestyle-config/shared/src/main/scala/config.scala
@@ -1,4 +1,4 @@
-package io.freestyle
+package freestyle
 
 import scala.concurrent.duration._
