+++ freestyle-effects/shared/src/main/scala/effects/option.scala
@@ -1,6 +1,6 @@
-package io.freestyle.effects
+package freestyle.effects
 
-import io.freestyle._
+import freestyle._
 import cats.MonadFilter
 
 object option {
