+++ freestyle-effects/shared/src/main/scala/effects/reader.scala
@@ -1,6 +1,6 @@
-package io.freestyle.effects
+package freestyle.effects
 
-import io.freestyle._
+import freestyle._
 import cats.{Eval, Monad, MonadReader}
 import cats.arrow.FunctionK
