+++ freestyle-logging/js/src/test/scala/io/freestyle/LoggingTests.scala
@@ -1,9 +1,9 @@
-package io.freestyle
+package freestyle
 
 import cats.Applicative
 import cats.instances.future._
-import io.freestyle.implicits._
-import io.freestyle.loggingJS.implicits._
+import freestyle.implicits._
+import freestyle.loggingJS.implicits._
 import org.scalatest._
 
 import scala.concurrent._
