+++ freestyle-logging/jvm/src/main/scala/io/freestyle/loggingJVM.scala
@@ -1,7 +1,7 @@
-package io.freestyle
+package freestyle
 
 import cats.MonadError
-import io.freestyle.logging._
+import freestyle.logging._
 import journal._
 
 object loggingJVM {
