+++ freestyle/shared/src/main/scala/freestyle/Lift.scala
@@ -1,4 +1,4 @@
-package io.freestyle
+package freestyle
 
 trait FreeSLift[F[_], G[_]] {
   def liftFS[A](ga: G[A]): FreeS[F, A] = liftFSPar[A](ga).freeS
