--- freestyle/shared/src/main/scala/io/freestyle/Implicits.scala
@@ -1,3 +1,0 @@
-package io.freestyle
-
-object implicits extends Interpreters
