+++ project/plugins.sbt
@@ -1,7 +1,7 @@
-addSbtPlugin("com.fortysevendeg" % "sbt-catalysts-extras" % "0.1.0")
-
-addSbtPlugin("com.fortysevendeg"  % "sbt-microsites" % "0.4.0")
+addSbtPlugin("com.fortysevendeg" % "sbt-catalysts-extras" % "0.1.2")
 
 addSbtPlugin("com.geirsson" % "sbt-scalafmt" % "0.4.10")
 
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.3.5")
+
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.14")
